# server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/videos": {
      "post": {
        "operationId": "VideosController_createVideo",
        "summary": "Create Video",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateVideoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The video has been successfully created."
          },
          "400": {
            "description": "Bad request."
          }
        },
        "tags": [
          "videos"
        ]
      },
      "get": {
        "operationId": "VideosController_findAllVideos",
        "summary": "Get All Videos",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all videos."
          }
        },
        "tags": [
          "videos"
        ]
      }
    },
    "/api/videos/{id}": {
      "get": {
        "operationId": "VideosController_findVideo",
        "summary": "Get Video by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Video ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The video has been successfully retrieved."
          },
          "400": {
            "description": "Invalid ID format."
          },
          "404": {
            "description": "Video not found."
          }
        },
        "tags": [
          "videos"
        ]
      },
      "patch": {
        "operationId": "VideosController_updateVideo",
        "summary": "Update video",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Video ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVideoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The video has been successfully updated."
          },
          "400": {
            "description": "Bad request."
          },
          "404": {
            "description": "Video not found."
          }
        },
        "tags": [
          "videos"
        ]
      },
      "delete": {
        "operationId": "VideosController_deleteVideo",
        "summary": "Delete video with content",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Video ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The video has been successfully deleted."
          },
          "400": {
            "description": "Invalid ID format."
          },
          "404": {
            "description": "Video not found."
          }
        },
        "tags": [
          "videos"
        ]
      }
    },
    "/api/videos/{videoId}/comments": {
      "post": {
        "operationId": "VideoCommentsController_createComment",
        "summary": "Create comment",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "Video ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The comment has been successfully created."
          },
          "400": {
            "description": "Bad request."
          },
          "404": {
            "description": "Video not found."
          }
        },
        "tags": [
          "comments"
        ]
      },
      "get": {
        "operationId": "VideoCommentsController_getComments",
        "summary": "Get all video comments",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "Video ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The comment has been successfully updated."
          },
          "400": {
            "description": "Invalid ID format."
          }
        },
        "tags": [
          "comments"
        ]
      }
    },
    "/api/videos/{videoId}/comments/{commentId}": {
      "patch": {
        "operationId": "VideoCommentsController_updateComment",
        "summary": "Update a comment",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "Video ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "description": "Comment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The comment has been successfully updated."
          },
          "400": {
            "description": "Invalid ID format."
          },
          "404": {
            "description": "Video or comment not found."
          }
        },
        "tags": [
          "comments"
        ]
      },
      "delete": {
        "operationId": "VideoCommentsController_deleteComment",
        "summary": "Delete a comment",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "Video ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "description": "Comment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The comment has been successfully deleted."
          },
          "400": {
            "description": "Invalid ID format."
          },
          "404": {
            "description": "Video or comment not found."
          }
        },
        "tags": [
          "comments"
        ]
      },
      "get": {
        "operationId": "VideoCommentsController_getComment",
        "summary": "Get comment by ID",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "Video ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "description": "Comment ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The comment has been successfully retrieved."
          },
          "400": {
            "description": "Invalid ID format."
          },
          "404": {
            "description": "Video or comment not found."
          }
        },
        "tags": [
          "comments"
        ]
      }
    },
    "/api/videos/{videoId}/comments/{commentId}/likes": {
      "post": {
        "operationId": "LikesController_likeComment",
        "summary": "Like a comment",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "ID of the video",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "description": "ID of the comment to like",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLikeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The comment has been successfully liked."
          },
          "400": {
            "description": "Invalid ID format."
          },
          "404": {
            "description": "Video or comment not found."
          }
        },
        "tags": [
          "likes"
        ]
      },
      "get": {
        "operationId": "LikesController_getCommentLikes",
        "summary": "Get likes of a comment",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "ID of the video",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "description": "ID of the comment to get like",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved like list."
          },
          "400": {
            "description": "Invalid ID format."
          },
          "404": {
            "description": "Video or comment not found."
          }
        },
        "tags": [
          "likes"
        ]
      }
    },
    "/api/videos/{videoId}/comments/{commentId}/likes/{likeId}": {
      "delete": {
        "operationId": "LikesController_unlikeComment",
        "summary": "Unlike a comment",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "ID of the video",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "description": "ID of the comment to unlike",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "likeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The comment has been successfully unliked."
          },
          "400": {
            "description": "Invalid ID format."
          },
          "404": {
            "description": "Video or comment not found or not liked by the user."
          }
        },
        "tags": [
          "likes"
        ]
      }
    },
    "/api/videos/{videoId}/comments/{commentId}/likes/count": {
      "get": {
        "operationId": "LikesController_getCommentLikeCount",
        "summary": "Get like count for a comment",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "ID of the video",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "description": "ID of the comment to get like count",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved like count."
          },
          "400": {
            "description": "Invalid ID format."
          },
          "404": {
            "description": "Video or comment not found."
          }
        },
        "tags": [
          "likes"
        ]
      }
    },
    "/api/videos/{videoId}/comments/{commentId}/likes/{userName}": {
      "get": {
        "operationId": "LikesController_getCommentLikesByUsername",
        "summary": "Get like of comment by username",
        "parameters": [
          {
            "name": "videoId",
            "required": true,
            "in": "path",
            "description": "ID of the video",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "description": "ID of the comment",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userName",
            "required": true,
            "in": "path",
            "description": "Name of the user to get like",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved like."
          },
          "400": {
            "description": "Invalid ID format."
          },
          "404": {
            "description": "Video or comment not found."
          }
        },
        "tags": [
          "likes"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "summary": "Register a new user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully registered."
          },
          "400": {
            "description": "Bad request."
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "Login a user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User successfully logged in."
          },
          "400": {
            "description": "Invalid credentials."
          }
        }
      }
    }
  },
  "info": {
    "title": "Video sharing site",
    "description": "The API description of video sharing site",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "likes",
      "description": ""
    },
    {
      "name": "comments",
      "description": ""
    },
    {
      "name": "videos",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "CreateVideoDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the video",
            "example": "My Awesome Video"
          },
          "description": {
            "type": "string",
            "description": "A description of the video content",
            "example": "This video showcases my latest project"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateVideoDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "The title of the video",
            "example": "My Awesome Video"
          },
          "description": {
            "type": "string",
            "description": "A description of the video content",
            "example": "This video showcases my latest project"
          }
        }
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "The content of the comment",
            "example": "This is a great video!"
          },
          "userDisplayName": {
            "type": "string",
            "description": "The display name of the user who posted the comment",
            "example": "John Doe"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "The date and time when the comment was created",
            "example": "2023-04-21T12:00:00Z"
          },
          "video": {
            "type": "object",
            "description": "The ID of the video this comment belongs to",
            "example": "507f1f77bcf86cd799439011"
          }
        },
        "required": [
          "content",
          "createdAt",
          "video"
        ]
      },
      "UpdateCommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "The updated content of the comment",
            "example": "This is an updated comment"
          }
        }
      },
      "CreateLikeDto": {
        "type": "object",
        "properties": {
          "userDisplayName": {
            "type": "string",
            "description": "The display name of the user who is creating the like",
            "example": "John Doe"
          }
        },
        "required": [
          "userDisplayName"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {}
      },
      "LoginUserDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}